Renames "logger" utility directory into "logging"
