+++ src/rest.ts
@@ -15,11 +15,11 @@ import { delay } from './context/delay'
 import { fetch } from './context/fetch'
 
 /* Logging */
-import { prepareRequest } from './utils/logger/prepareRequest'
-import { prepareResponse } from './utils/logger/prepareResponse'
+import { prepareRequest } from './utils/logging/prepareRequest'
+import { prepareResponse } from './utils/logging/prepareResponse'
 import { getPublicUrlFromRequest } from './utils/request/getPublicUrlFromRequest'
-import { getTimestamp } from './utils/logger/getTimestamp'
-import { getStatusCodeColor } from './utils/logger/getStatusCodeColor'
+import { getTimestamp } from './utils/logging/getTimestamp'
+import { getStatusCodeColor } from './utils/logging/getStatusCodeColor'
 import { isStringEqual } from './utils/internal/isStringEqual'
 import { matchRequestUrl } from './utils/matching/matchRequest'
 import { resolveMask } from './utils/resolveMask'
