+++ src/utils/logging/getStatusCodeColor.test.ts
