+++ src/utils/logging/getStatusCodeColor.ts
