+++ src/utils/logging/prepareResponse.ts
